# cmd/server/domain/format_response.go
// Code generated by go-swagger; DO NOT EDIT.

package domain

import "context"

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// FormatResponse format response
//
// swagger:model FormatResponse
type FormatResponse struct {

	// code
	Code string `json:"code,omitempty"`
}

// Validate validates this format response
func (m *FormatResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this format response based on context it is used
func (m *FormatResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *FormatResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *FormatResponse) UnmarshalBinary(b []byte) error {
	var res FormatResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}