# cmd/server/domain/language_info.go
// Code generated by go-swagger; DO NOT EDIT.

package domain

import "context"

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// LanguageInfo language info
//
// swagger:model LanguageInfo
type LanguageInfo struct {

	// name
	Name string `json:"name,omitempty"`

	// version
	Version string `json:"version,omitempty"`
}

// Validate validates this language info
func (m *LanguageInfo) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this language info based on context it is used
func (m *LanguageInfo) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *LanguageInfo) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *LanguageInfo) UnmarshalBinary(b []byte) error {
	var res LanguageInfo
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}