# cmd/server/domain/run_request.go
// Code generated by go-swagger; DO NOT EDIT.

package domain

import "context"

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"encoding/json"

	"github.com/go-openapi/errors"
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
	"github.com/go-openapi/validate"
)

// RunRequest run request
//
// swagger:model RunRequest
type RunRequest struct {

	// code
	// Required: true
	Code *string `json:"code"`

	// id
	ID string `json:"id,omitempty"`

	// language
	// Required: true
	// Enum: ["golang","python","javascript","rust","cpp"]
	Language *string `json:"language"`
}

// Validate validates this run request
func (m *RunRequest) Validate(formats strfmt.Registry) error {
	var res []error

	if err := m.validateCode(formats); err != nil {
		res = append(res, err)
	}

	if err := m.validateLanguage(formats); err != nil {
		res = append(res, err)
	}

	if len(res) > 0 {
		return errors.CompositeValidationError(res...)
	}
	return nil
}

func (m *RunRequest) validateCode(formats strfmt.Registry) error {

	if err := validate.Required("code", "body", m.Code); err != nil {
		return err
	}

	return nil
}

var runRequestTypeLanguagePropEnum []any

func init() {
	var res []string
	if err := json.Unmarshal([]byte(`["golang","python","javascript","rust","cpp"]`), &res); err != nil {
		panic(err)
	}
	for _, v := range res {
		runRequestTypeLanguagePropEnum = append(runRequestTypeLanguagePropEnum, v)
	}
}

const (

	// RunRequestLanguageGolang captures enum value "golang"
	RunRequestLanguageGolang string = "golang"

	// RunRequestLanguagePython captures enum value "python"
	RunRequestLanguagePython string = "python"

	// RunRequestLanguageJavascript captures enum value "javascript"
	RunRequestLanguageJavascript string = "javascript"

	// RunRequestLanguageRust captures enum value "rust"
	RunRequestLanguageRust string = "rust"

	// RunRequestLanguageCpp captures enum value "cpp"
	RunRequestLanguageCpp string = "cpp"
)

// prop value enum
func (m *RunRequest) validateLanguageEnum(path, location string, value string) error {
	if err := validate.EnumCase(path, location, value, runRequestTypeLanguagePropEnum, true); err != nil {
		return err
	}
	return nil
}

func (m *RunRequest) validateLanguage(formats strfmt.Registry) error {

	if err := validate.Required("language", "body", m.Language); err != nil {
		return err
	}

	// value enum
	if err := m.validateLanguageEnum("language", "body", *m.Language); err != nil {
		return err
	}

	return nil
}

// ContextValidate validates this run request based on context it is used
func (m *RunRequest) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *RunRequest) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *RunRequest) UnmarshalBinary(b []byte) error {
	var res RunRequest
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}