# cmd/server/domain/run_response.go
// Code generated by go-swagger; DO NOT EDIT.

package domain

import "context"

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// RunResponse run response
//
// swagger:model RunResponse
type RunResponse struct {

	// id
	ID string `json:"id,omitempty"`

	// output
	Output string `json:"output,omitempty"`
}

// Validate validates this run response
func (m *RunResponse) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this run response based on context it is used
func (m *RunResponse) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *RunResponse) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *RunResponse) UnmarshalBinary(b []byte) error {
	var res RunResponse
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}