# cmd/server/domain/snippet.go
// Code generated by go-swagger; DO NOT EDIT.

package domain

import "context"

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// Snippet snippet
//
// swagger:model Snippet
type Snippet struct {

	// code
	Code string `json:"code,omitempty"`

	// id
	ID string `json:"id,omitempty"`

	// language
	Language string `json:"language,omitempty"`

	// output
	Output string `json:"output,omitempty"`
}

// Validate validates this snippet
func (m *Snippet) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this snippet based on context it is used
func (m *Snippet) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *Snippet) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *Snippet) UnmarshalBinary(b []byte) error {
	var res Snippet
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}